feat(metadata): implement load_metadata function and restructure metadata loading
